fix(MemoryOrb): Enhance glassmorphism and interactivity
Refine the visual appearance of MemoryOrb components to better embody a "glass orb" aesthetic. This includes adjusting borders, shadows, and background for a more translucent and responsive feel on hover.

The `getFrontAndCenterPos` function has been added to calculate the spherical coordinates of the "front" of the view based on the current world rotation. This will be used to position new memories more intelligently.